src/requirements.txt
aiogram>=3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
undetected-chromedriver==3.5.3
selenium==4.15.2
